Create README.md
This is my first project on github so please bear with me ...

I started this project because we've had many files in an old office format on the file server in our company and we stopped accepting mails with attachments in these formats due to the high risk of getting malware through office files with macros.
As our users sent out mails with attachments in these formats due to fact that they are lazy and using tons of old files as a poor mans template (copy an old file, rename it, change it, send it) I decided to develop a small tool to convert the files into the new OpenXML format.

The tool let you scan a directory including all sub directories for files with an old format (*.xls & *.doc) and will convert them automatically by opening an Excel or Word instance in the background an then save the file in new OpenXML format. The tool detects files containing macros and try to handle them accordingly. It also detects files with an extension .xls or .doc but containing no Office formats (they will be skipped and not converted).

Maybe this tool is useful for somebody else in the same situation we had...